Pipeline debugging (#26)
* Updated ingest process to work with the open data process

After measurements are ingested we update the open data log to trigger
an export

* Updated the load metadata method and logging

Fixed bug that was preventing fetchlogs files from being
finished. Keys were being pulled down, then cleaned and then used to
reference the fetchlog again after the key was changed.

Also added logging and removed the print statements

* Cleaning up the load_measurements method

Broke the method up into two methods to make it easier to test

* Reverted the load_measurements_db query

Flipping to asc makes it easier to load old files that have not been
processed and therefor avoid overlapping with the lambda

* Fixed one sensor_node insert bug

Also added the times to the rejects inserts

* Added a manual method to the handler

The goal is to be able to test the lambda under different conditions

* More timing updates

* Some tweeks to the handler method

* Added quick patch to deal with realtime errors

Patch catches and logs error as well as marking the file as completed
but with an error

* Added group-by to the insert sensors query

This is to deal with the issue where we have duplicates in the data
file being inserted

* Simplifying the realtime ingest process

Putting all of the ingest SQL into one file and added metrics that can
be logged out and tracked.

* Updated the fetch method to try and get a better error message

Removed the try/catch so that we did not have nested exceptions

* Cleaned up the ingest sql files, added utility methods

Combined them and added some tracking metrics. Also broke up some of
the functions and added utility methods to make it easier to diagnose
different issues.

* Added fetchlogs_id to the ingest process

Added for tracing bad data back to the source. Goes all the way to the
    rejects table.

* Added fetchlogs_id to the meas ingest method

* Added try blocks at the handler level

* Updated the lcs methods to include fetchlogs_id

* Added some more check methods

* Created better check method

Method allows for check by id, fix and dryrun

* Tweaks to the check code

- Added more options
- Updated the message saved to the fetchlog to show it was fixed

* Started SOP file

* Updated the SOP

* Removed update to last_modified

We dont need that to be changed

* Added more rejects functionality

* Updated the check_handler method

Wanted to make it more like the check method and include local file
upload method

* Some cleanup

* Added some more arguments to the MAINTENANCE table

* Added help to the check.py file

Co-authored-by: Christian Parker <cparker@talloaks-environmental.com>