Removed jq from the v1/latest/{id} endpoint
This was to determine what it would take to remove our dependence on
jq. The key changes here are in the db class where I added support for
json. However, this added support will break the old
`fetchOpenAQResult` method and so that needed to be modified to handle
the changes. From there its just a matter of using sql to reshape the
data instead of jq. I also changed the way in which we are getting the
total, however that is just a quick hack to make this test work.